pyproject.toml
[project]
name = "agent_core_deep_research"
version = "0.1.0"
description = "Build a deep research agent from scratch"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain_community>=0.4",
    "langchain_tavily>=0.2.12",
    "langchain_mcp_adapters>=0.1.11",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
    "tavily-python>=0.5.0",
    "bedrock-agentcore>=1.0.5",
    "pytest>=9.0.0",
    "bedrock-agentcore-starter-toolkit>=0.1.32",
    "strands-agents>=1.15.0",
    "langchain-aws>=1.0.0",
    "python-dotenv>=1.1.1",
    "botocore>=1.40.69",
    "aws-cdk-lib>=2.223.0",
    "constructs>=10.4.3",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deep_research", "agent_core"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
